Extract Slack pending message handling
